minor fixes plus renamings needed in a few components
