+++ components/layout/header/header-about/index.js
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux';
 import * as actions from '../header-actions';
 
-import HeaderAboutComponent from './header-about-component';
+import HeaderAboutComponent from './component';
 
 export default connect(
   state => ({
