+++ components/layout/header/header-data/index.js
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux';
 import * as actions from '../header-actions';
 
-import HeaderDataComponent from './header-data-component';
+import HeaderDataComponent from './component';
 
 export default connect(
   state => ({
