+++ components/layout/header/header-user/index.js
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux';
 import * as actions from '../header-actions';
 
-import HeaderUserComponent from './header-user-component';
+import HeaderUserComponent from './component';
 
 export default connect(
   state => ({
